// src/api/models/questionConstraint.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * GradeFlow API
 * Thin FastAPI wrapper over GradeFlow Engine
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionConstraintType } from './questionConstraintType';

export interface QuestionConstraint {
  type: QuestionConstraintType;
  source: string;
  target: string;
}